fix statFilter logic, add api directory
