+++ src/api/const.js
@@ -1,0 +1,15 @@
+export const AUTHORIZATION = `Basic t2i4titi5ky7if3`;
+
+export const END_POINT = `https://13.ecmascript.pages.academy/cinemaddict`;
+
+export const Method = {
+  GET: `GET`,
+  PUT: `PUT`,
+  POST: `POST`,
+  DELETE: `DELETE`
+};
+
+export const SuccessHHTPStatusRange = {
+  MIN: 200,
+  MAX: 299
+};
