+++ src/const.js
@@ -6,10 +6,6 @@ export const MAX_DESCRIPTION_SYMBOLS = 140;
 
 export const BAR_HEIGHT = 50;
 
-export const AUTHORIZATION = `Basic t2i4titi5ky7if3`;
-
-export const END_POINT = `https://13.ecmascript.pages.academy/cinemaddict`;
-
 export const ERROR_COMMENTS_UPLOAD = `failed to load`;
 
 export const COMMENT_DELETING = `deleting`;
@@ -65,16 +61,4 @@ export const StatFilterType = {
   WEEK: `week`,
   MONTH: `month`,
   YEAR: `year`
-};
-
-export const Method = {
-  GET: `GET`,
-  PUT: `PUT`,
-  POST: `POST`,
-  DELETE: `DELETE`
-};
-
-export const SuccessHHTPStatusRange = {
-  MIN: 200,
-  MAX: 299
 };
